+++ Content/Items/Weapons/Summon/Fumigator/FumigatorMinion.cs
@@ -46,6 +46,7 @@ namespace SuperfluityTwo.Content.Items.Weapons.Summon.Fumigator
 			Projectile.DamageType = DamageClass.Summon; // Declares the damage type (needed for it to deal damage)
 			Projectile.minionSlots = 1f; // Amount of slots this minion occupies from the total minion slots available to the player (more on that later)
 			Projectile.penetrate = -1; // Needed so the minion doesn't despawn on collision with enemies or tiles
+			Projectile.netImportant = true;
 		}
 
 		// Here you can decide if your minion breaks things like grass or pots