+++ Content/Items/Weapons/Summon/SafetyLamp/SafetyLampItem.cs
@@ -1,13 +1,15 @@
+using System;
 using Microsoft.Xna.Framework;
+using SuperfluityTwo.Content.Items.Weapons.Summon.Asterism;
 using SuperfluityTwo.Content.Projectiles;
 using Terraria;
 using Terraria.DataStructures;
 using Terraria.ID;
 using Terraria.ModLoader;
 
-namespace SuperfluityTwo.Content.Items.Weapons.Summon.Asterism
+namespace SuperfluityTwo.Content.Items.Weapons.Summon.SafetyLamp
 {
-	public class AsterismItem : GlowItem
+	public class SafetyLampItem : GlowItem
 	{
         public override void SetStaticDefaults() {
 			ItemID.Sets.GamepadWholeScreenUseRange[Item.type] = true;
@@ -18,39 +20,34 @@ namespace SuperfluityTwo.Content.Items.Weapons.Summon.Asterism
 
         public override void SetDefaults()
 		{
-			Item.width = 20;
-			Item.height = 20;
+			Item.width = 14;
+			Item.height = 30;
 			Item.mana = 12;
-			Item.useStyle = ItemUseStyleID.Swing;
+			Item.holdStyle = ItemHoldStyleID.HoldLamp;
+			Item.useStyle = ItemUseStyleID.RaiseLamp;
 			Item.UseSound = SoundID.Item4;
 			Item.useAnimation = 26;
 			Item.useTime = 26;
 			Item.noMelee = true;
-			//Item.crit = 0;
-			Item.damage = 18;
+			Item.scale = 0.7f;
+			Item.damage = 4;
 			Item.DamageType = DamageClass.Summon;
-			Item.knockBack = 7.5f;
-			Item.value = Item.sellPrice(gold: 2);
+			Item.knockBack = 1.5f;
+			Item.value = Item.sellPrice(gold: 3);
 			Item.rare = ItemRarityID.Green;
 			Item.useTurn = false;
-			Item.buffType = ModContent.BuffType<AsterismBuff>();
-			Item.shoot = ModContent.ProjectileType<AsterismMinion>();
+			Item.buffType = ModContent.BuffType<SafetyLampBuff>();
+			Item.shoot = ModContent.ProjectileType<SafetyLampCounter>();
 		}
 
         public override void AddRecipes()
 		{
 			CreateRecipe()
-				.AddIngredient(ItemID.MagicMissile)
-				.AddIngredient(ItemID.FallenStar, 20)
-				.AddTile(TileID.ShimmerMonolith)
+				.AddIngredient(ItemID.StarinaBottle)
+				.AddIngredient(ItemID.Lens)
+				.AddRecipeGroup(RecipeGroupID.IronBar, 5)
+				.AddTile(TileID.Anvils)
 				.Register();
-		}
-
-		public override void ModifyShootStats(Player player, ref Vector2 position, ref Vector2 velocity, ref int type, ref int damage, ref float knockback)
-		{
-			// Here you can change where the minion is spawned. Most vanilla minions spawn at the cursor position
-			position = Main.MouseWorld;
-			player.LimitPointToPlayerReachableArea(ref position);
 		}
 
 		public override bool Shoot(Player player, EntitySource_ItemUse_WithAmmo source, Vector2 position, Vector2 velocity, int type, int damage, float knockback) {
@@ -64,6 +61,8 @@ namespace SuperfluityTwo.Content.Items.Weapons.Summon.Asterism
 
 			// Since we spawned the projectile manually already, we do not need the game to spawn it for ourselves anymore, so return false
 			*/
+			player.GetModPlayer<SafetyLampPlayer>().highestOriginalDamage = Math.Max(Item.damage, player.GetModPlayer<SafetyLampPlayer>().highestOriginalDamage);
+			player.GetModPlayer<SafetyLampPlayer>().highestKnockback = Math.Max(Item.knockBack, player.GetModPlayer<SafetyLampPlayer>().highestKnockback);
 			return true;
 		}